Add option to choose offset/limit for substrings.
This offset/limit is based on the list of objects
that get returned from the FM-index,
not on the amount of triples they are matched with.
To get all the results, simply set the limit to 0.
The value returned from the function if the total number of results.
The num_count output variable is the actual length of the occs array,
based on the offset and limit values.
Setting deduplicate to true and using offset/limit
can lead to incomplete results.

Also updated naiveComplete to allow for offset and limit parameters.